backend/requirements.txt
fastapi>=0.115.0
anyio>=4.0.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
pydantic[email]>=2.10.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
psycopg[binary]>=3.2.0
alembic>=1.14.0
passlib[argon2]>=1.7.4
argon2-cffi>=23.0.0
python-jose[cryptography]>=3.3.0
email-validator>=2.2.0
anthropic>=0.40.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
aiosqlite>=0.20.0
